@@ -1,6 +1,6 @@
 {
   "name": "node-cqrs-lib",
-  "version": "0.5",
+  "version": "0.6",
   "description": "Library of base classes for Node CQRS",
   "main": "index.js",
   "scripts": {
@@ -22,50 +22,47 @@
   "dependencies": {
     "node-fetch": "^1.6.3",
     "reflect-metadata": "^0.1.10",
-    "uuid": "~3.0.1"
+    "uuid": "3.0.1"
   },
   "devDependencies": {
-    "babel-cli": "~6.18.0",
-    "babel-core": "~6.21.0",
-    "babel-eslint": "~7.1.1",
-    "babel-plugin-add-module-exports": "~0.2.1",
-    "babel-plugin-transform-async-to-generator": "~6.16.0",
-    "babel-plugin-transform-class-properties": "~6.19.0",
-    "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-plugin-transform-es2015-destructuring": "~6.19.0",
-    "babel-plugin-transform-object-rest-spread": "~6.20.2",
-    "babel-plugin-transform-runtime": "~6.15.0",
-    "babel-polyfill": "~6.20.0",
-    "babel-preset-es2015": "~6.18.0",
-    "babel-preset-es2017": "~6.16.0",
-    "babel-runtime": "~6.20.0",
-    "chai": "~3.5.0",
-    "chai-as-promised": "^6.0.0",
-    "chai-datetime": "^1.5.0",
-    "del": "~2.2.2",
-    "eslint": "~3.13.1",
-    "gulp": "~3.9.1",
-    "gulp-babel": "~6.1.2",
-    "gulp-cached": "~1.1.0",
-    "gulp-changed": "~1.3.2",
-    "gulp-debug": "~3.0.0",
-    "gulp-eslint": "~3.0.1",
-    "gulp-if": "~2.0.2",
-    "gulp-istanbul": "~1.1.1",
-    "gulp-load-plugins": "~1.4.0",
-    "gulp-mocha": "~3.0.1",
-    "gulp-remember": "~0.3.1",
-    "gulp-size": "~2.1.0",
-    "gulp-sourcemaps": "~2.4.0",
-    "gulp-uglify": "~2.0.0",
-    "gulp-util": "~3.0.8",
-    "isparta": "~4.0.0",
-    "istanbul": "~0.4.5",
-    "jshint-stylish": "~2.2.1",
-    "mocha": "~3.2.0",
-    "node-mocks-http": "~1.5.4",
-    "proxyquire": "~1.7.10",
-    "sinon": "~1.17.7",
-    "supertest": "~2.0.0"
+    "@babel/cli": "7.0.0-beta.38",
+    "@babel/core": "7.0.0-beta.38",
+    "@babel/plugin-proposal-object-rest-spread": "7.0.0-beta.38",
+    "@babel/plugin-proposal-class-properties": "7.0.0-beta.38",
+    "@babel/plugin-proposal-decorators": "7.0.0-beta.38",
+    "@babel/plugin-proposal-optional-chaining": "7.0.0-beta.38",
+    "@babel/polyfill": "7.0.0-beta.38",
+    "@babel/preset-env": "7.0.0-beta.38",
+    "@babel/register": "7.0.0-beta.38",
+    "@babel/runtime": "7.0.0-beta.38",
+    "babel-istanbul": "0.12.2",
+    "babel-eslint": "8.2.1",
+    "chai": "4.1.2",
+    "chai-as-promised": "7.1.1",
+    "chai-datetime": "1.5.0",
+    "del": "3.0.0",
+    "eslint": "4.15.0",
+    "gulp": "4.0.0",
+    "gulp-babel": "8.0.0-beta.0",
+    "gulp-cached": "1.1.1",
+    "gulp-changed": "3.2.0",
+    "gulp-debug": "3.2.0",
+    "gulp-eslint": "4.0.1",
+    "gulp-if": "2.0.2",
+    "gulp-istanbul": "1.1.3",
+    "gulp-load-plugins": "1.5.0",
+    "gulp-mocha": "5.0.0",
+    "gulp-remember": "1.0.1",
+    "gulp-size": "3.0.0",
+    "gulp-sourcemaps": "2.6.3",
+    "gulp-uglify": "3.0.0",
+    "gulp-util": "3.0.8",
+    "isparta": "4.0.0",
+    "jshint-stylish": "2.2.1",
+    "mocha": "5.0.0",
+    "node-mocks-http": "1.6.6",
+    "proxyquire": "1.8.0",
+    "sinon": "4.1.6",
+    "supertest": "3.0.0"
   }
 }
