@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),Aggregate=function(){function e(r){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:[];(0,_classCallCheck3.default)(this,e),this.id=r,this.version=0,this.applyEvents(t)}return(0,_createClass3.default)(e,[{key:"applyEvents",value:function(e){var r=this;return(Array.isArray(e)?e:[e]).map(function(e){if(e.sequenceNumber&&e.sequenceNumber<=r.version)throw new Error("Event came out of sequence");return e.update(r),e.sequenceNumber=e.sequenceNumber||r.version+1,e.aggregateId=e.aggregateId||r.id,r.version=e.sequenceNumber,e})}}]),e}();exports.default=Aggregate,module.exports=exports.default;
+"use strict";function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var Aggregate=function(){function e(r){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:[];_classCallCheck(this,e),this.id=r,this.version=0,this.applyEvents(t)}return _createClass(e,[{key:"applyEvents",value:function(e){var r=this;return(Array.isArray(e)?e:[e]).map(function(e){if(e.sequenceNumber&&e.sequenceNumber<=r.version)throw new Error("Event came out of sequence");return e.update(r),e.sequenceNumber=e.sequenceNumber||r.version+1,e.aggregateId=e.aggregateId||r.id,r.version=e.sequenceNumber,e})}}]),e}();exports.default=Aggregate;
 //# sourceMappingURL=maps/aggregate.js.map
