@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function asArray(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return Array.isArray(e)?e:[e]}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),AggregateCommandHandler=function(){function e(r){var t=this;(0,_classCallCheck3.default)(this,e),this.handle=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r,a){var n,u;return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return n=new t.Command(r),e.next=3,n.validate();case 3:return e.next=5,t.authorize(n,a);case 5:return e.next=7,t.validate(n,a);case 7:return e.t0=asArray,e.next=10,t.execute(n,a);case 10:return e.t1=e.sent,u=(0,e.t0)(e.t1),r.$scheduler=n.$scheduler,u.forEach(function(e){e.actor=n.$identity&&n.$identity.userId,e.position=n.$position}),e.abrupt("return",a.applyEvents(u));case 15:case"end":return e.stop()}},e,t)}));return function(r,t){return e.apply(this,arguments)}}(),this.execute=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r,a){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,t)}));return function(r,t){return e.apply(this,arguments)}}(),this.authorize=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r,a){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,t)}));return function(r,t){return e.apply(this,arguments)}}(),this.handleDeliveryError=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:console.log("delivery failure",r),r.cancel();case 2:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.Command=r}return(0,_createClass3.default)(e,[{key:"validate",value:function(e,r){}}]),e}();exports.default=AggregateCommandHandler,module.exports=exports.default;
+"use strict";function _asyncToGenerator(e){return function(){var r=this,t=arguments;return new Promise(function(n,a){var i=e.apply(r,t);function o(e,r){try{var t=i[e](r),o=t.value}catch(e){return void a(e)}t.done?n(o):Promise.resolve(o).then(u,c)}function u(e){o("next",e)}function c(e){o("throw",e)}u()})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var AggregateCommandHandler=function(){function e(r){var t,n,a,i,o=this;_classCallCheck(this,e),Object.defineProperty(this,"handle",{configurable:!0,enumerable:!0,writable:!0,value:(t=_asyncToGenerator(regeneratorRuntime.mark(function e(r,t){var n,a;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return n=new o.Command(r),e.next=3,n.validate();case 3:return e.next=5,o.authorize(n,t);case 5:return e.next=7,o.validate(n,t);case 7:return e.t0=asArray,e.next=10,o.execute(n,t);case 10:return e.t1=e.sent,a=(0,e.t0)(e.t1),r.$scheduler=n.$scheduler,a.forEach(function(e){e.actor=n.$identity&&n.$identity.userId,e.position=n.$position}),e.abrupt("return",t.applyEvents(a));case 15:case"end":return e.stop()}},e,this)})),function(e,r){return t.apply(this,arguments)})}),Object.defineProperty(this,"execute",{configurable:!0,enumerable:!0,writable:!0,value:(n=_asyncToGenerator(regeneratorRuntime.mark(function e(r,t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,this)})),function(e,r){return n.apply(this,arguments)})}),Object.defineProperty(this,"authorize",{configurable:!0,enumerable:!0,writable:!0,value:(a=_asyncToGenerator(regeneratorRuntime.mark(function e(r,t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,this)})),function(e,r){return a.apply(this,arguments)})}),Object.defineProperty(this,"handleDeliveryError",{configurable:!0,enumerable:!0,writable:!0,value:(i=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:console.log("delivery failure",r),r.cancel();case 2:case"end":return e.stop()}},e,this)})),function(e){return i.apply(this,arguments)})}),this.Command=r}return _createClass(e,[{key:"validate",value:function(e,r){}}]),e}();function asArray(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return Array.isArray(e)?e:[e]}exports.default=AggregateCommandHandler;
 //# sourceMappingURL=maps/aggregateCommandHandler.js.map
