@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),_identity=require("./identity"),_identity2=_interopRequireDefault(_identity),Command=function(){function e(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};(0,_classCallCheck3.default)(this,e),this.validate=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,this)})),this.$position=t.$position,this.$identity=new _identity2.default(t.$identity)||_identity2.default.anonymous,this.$scheduler=t.$scheduler||{}}return(0,_createClass3.default)(e,[{key:"retry",value:function(e){var t=e.due,r=e.seconds;t||(t=new Date,t.setSeconds(t.getSeconds()+r)),this.$scheduler.due=t}}]),e}();exports.default=Command,module.exports=exports.default;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _identity=_interopRequireDefault(require("./identity"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _asyncToGenerator(e){return function(){var t=this,n=arguments;return new Promise(function(r,i){var o=e.apply(t,n);function a(e,t){try{var n=o[e](t),a=n.value}catch(e){return void i(e)}n.done?r(a):Promise.resolve(a).then(u,s)}function u(e){a("next",e)}function s(e){a("throw",e)}u()})}}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function _createClass(e,t,n){return t&&_defineProperties(e.prototype,t),n&&_defineProperties(e,n),e}var Command=function(){function e(){var t,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};_classCallCheck(this,e),Object.defineProperty(this,"validate",{configurable:!0,enumerable:!0,writable:!0,value:(t=_asyncToGenerator(regeneratorRuntime.mark(function e(){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,this)})),function(){return t.apply(this,arguments)})}),this.$position=n.$position,this.$identity=new _identity.default(n.$identity)||_identity.default.anonymous,this.$scheduler=n.$scheduler||{}}return _createClass(e,[{key:"retry",value:function(e){var t=e.due,n=e.seconds;t||(t=new Date).setSeconds(t.getSeconds()+n),this.$scheduler.due=t}}]),e}();exports.default=Command;
 //# sourceMappingURL=maps/command.js.map
