@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),CommandFailure=function(){function e(a){var l=a.error,r=a.command,t=a.aggregate;(0,_classCallCheck3.default)(this,e),this.error=l,this.command=r,this.aggregate=t}return(0,_createClass3.default)(e,[{key:"retry",value:function(e){var a=e.due,l=void 0===a?null:a,r=e.seconds,t=void 0===r?null:r;l||(l=new Date,l.setSeconds(l.getSeconds()+t)),this.command.$scheduler.due=l}},{key:"cancel",value:function(){this.cancelled=!0}}]),e}();exports.default=CommandFailure,module.exports=exports.default;
+"use strict";function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var n=0;n<r.length;n++){var t=r[n];t.enumerable=t.enumerable||!1,t.configurable=!0,"value"in t&&(t.writable=!0),Object.defineProperty(e,t.key,t)}}function _createClass(e,r,n){return r&&_defineProperties(e.prototype,r),n&&_defineProperties(e,n),e}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var CommandFailure=function(){function e(r){var n=r.error,t=r.command,a=r.aggregate;_classCallCheck(this,e),this.error=n,this.command=t,this.aggregate=a}return _createClass(e,[{key:"retry",value:function(e){var r=e.due,n=void 0===r?null:r,t=e.seconds,a=void 0===t?null:t;n||(n=new Date).setSeconds(n.getSeconds()+a),this.command.$scheduler.due=n}},{key:"cancel",value:function(){this.cancelled=!0}}]),e}();exports.default=CommandFailure;
 //# sourceMappingURL=../maps/commandScheduling/commandFailure.js.map
