@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_stringify=require("babel-runtime/core-js/json/stringify"),_stringify2=_interopRequireDefault(_stringify),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),_url=require("url"),_url2=_interopRequireDefault(_url),_identity=require("../identity"),_identity2=_interopRequireDefault(_identity),RestCommandDeliverer=function(){function e(r,t,a){var n=this;(0,_classCallCheck3.default)(this,e),this.deliver=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){var t,a,i=r.service,s=r.target,o=r.command;return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=n.formatCommandUrl({service:i,target:s,command:o.type}),e.next=3,n.getSystemIdentity();case 3:return a=e.sent,e.next=6,n.fetch(t,{method:"POST",headers:{"Content-Type":"application/json",authorization:a,position:o.$position&&new Buffer((0,_stringify2.default)(o.$position)).toString("base64")},body:(0,_stringify2.default)(o)});case 6:return e.abrupt("return",e.sent.json());case 7:case"end":return e.stop()}},e,n)}));return function(r){return e.apply(this,arguments)}}(),this.baseUrl=_url2.default.parse(r+(r.endsWith("/")?"":"/")),this.fetch=t,this.getSystemIdentity=a}return(0,_createClass3.default)(e,[{key:"formatCommandUrl",value:function(e){var r=e.service,t=e.target,a=e.command;return _url2.default.format({protocol:this.baseUrl.protocol,host:this.baseUrl.host,pathname:""+this.baseUrl.pathname+(r?r+"/":"")+(t?t+"/":"")+a})}}]),e}();exports.default=RestCommandDeliverer,module.exports=exports.default;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _url=_interopRequireDefault(require("url"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _asyncToGenerator(e){return function(){var t=this,r=arguments;return new Promise(function(n,a){var o=e.apply(t,r);function i(e,t){try{var r=o[e](t),i=r.value}catch(e){return void a(e)}r.done?n(i):Promise.resolve(i).then(s,u)}function s(e){i("next",e)}function u(e){i("throw",e)}s()})}}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,t,r){return t&&_defineProperties(e.prototype,t),r&&_defineProperties(e,r),e}var RestCommandDeliverer=function(){function e(t,r,n){var a,o=this;_classCallCheck(this,e),Object.defineProperty(this,"deliver",{configurable:!0,enumerable:!0,writable:!0,value:(a=_asyncToGenerator(regeneratorRuntime.mark(function e(t){var r,n,a,i,s;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return r=t.service,n=t.target,a=t.command,i=o.formatCommandUrl({service:r,target:n,command:a.type}),e.next=4,o.getSystemIdentity();case 4:return s=e.sent,e.next=7,o.fetch(i,{method:"POST",headers:{"Content-Type":"application/json",authorization:s,position:a.$position&&new Buffer(JSON.stringify(a.$position)).toString("base64")},body:JSON.stringify(a)});case 7:return e.abrupt("return",e.sent.json());case 8:case"end":return e.stop()}},e,this)})),function(e){return a.apply(this,arguments)})}),this.baseUrl=_url.default.parse(t+(t.endsWith("/")?"":"/")),this.fetch=r,this.getSystemIdentity=n}return _createClass(e,[{key:"formatCommandUrl",value:function(e){var t=e.service,r=e.target,n=e.command;return _url.default.format({protocol:this.baseUrl.protocol,host:this.baseUrl.host,pathname:"".concat(this.baseUrl.pathname).concat(t?t+"/":"").concat(r?r+"/":"").concat(n)})}}]),e}();exports.default=RestCommandDeliverer;
 //# sourceMappingURL=../maps/commandScheduling/restCommandDeliverer.js.map
