@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),Event=function(){function e(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};(0,_classCallCheck3.default)(this,e),this.aggregate=t.aggregate,this.aggregateId=t.aggregateId,this.sequenceNumber=t.sequenceNumber,this.timestamp=t.timestamp,this.actor=t.actor,this.position=t.position}return(0,_createClass3.default)(e,[{key:"update",value:function(e){}}]),e}();exports.default=Event,module.exports=exports.default;
+"use strict";function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,t){for(var r=0;r<t.length;r++){var a=t[r];a.enumerable=a.enumerable||!1,a.configurable=!0,"value"in a&&(a.writable=!0),Object.defineProperty(e,a.key,a)}}function _createClass(e,t,r){return t&&_defineProperties(e.prototype,t),r&&_defineProperties(e,r),e}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var Event=function(){function e(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};_classCallCheck(this,e),this.aggregate=t.aggregate,this.aggregateId=t.aggregateId,this.sequenceNumber=t.sequenceNumber,this.timestamp=t.timestamp,this.actor=t.actor,this.position=t.position}return _createClass(e,[{key:"update",value:function(e){}}]),e}();exports.default=Event;
 //# sourceMappingURL=maps/event.js.map
