@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _promise=require("babel-runtime/core-js/promise"),_promise2=_interopRequireDefault(_promise),_regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_createClass2=require("babel-runtime/helpers/createClass"),_createClass3=_interopRequireDefault(_createClass2),_v=require("uuid/v4"),_v2=_interopRequireDefault(_v),Repository=function(){function e(r){var t=this,n=r.eventStore,a=r.aggregate;(0,_classCallCheck3.default)(this,e),this.get=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){var n;return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.eventStore.getEvents(r);case 2:if(n=e.sent,n.length){e.next=5;break}return e.abrupt("return",null);case 5:return e.abrupt("return",new t.Ctor(r,n));case 6:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.create=function(e){return new t.Ctor(e||(0,_v2.default)())},this.record=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){var n;return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.eventStore.record(r);case 2:return n=e.sent,e.next=5,t.notifySubscribers(n);case 5:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.notifySubscribers=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,_promise2.default.all(t.subscriptions.map(function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(n){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,n(r);case 3:return e.abrupt("return",e.sent);case 6:e.prev=6,e.t0=e.catch(0),console.log(e.t0);case 9:case"end":return e.stop()}},e,t,[[0,6]])}));return function(r){return e.apply(this,arguments)}}()));case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.getEvents=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.eventStore.getEvents(r);case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.eventStore=n,this.Ctor=a,this.subscriptions=[]}return(0,_createClass3.default)(e,[{key:"subscribe",value:function(e){this.subscriptions.push(e)}}]),e}();exports.default=Repository,module.exports=exports.default;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _v=_interopRequireDefault(require("uuid/v4"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _asyncToGenerator(e){return function(){var r=this,t=arguments;return new Promise(function(n,a){var u=e.apply(r,t);function o(e,r){try{var t=u[e](r),o=t.value}catch(e){return void a(e)}t.done?n(o):Promise.resolve(o).then(i,s)}function i(e){o("next",e)}function s(e){o("throw",e)}i()})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}function _defineProperties(e,r){for(var t=0;t<r.length;t++){var n=r[t];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function _createClass(e,r,t){return r&&_defineProperties(e.prototype,r),t&&_defineProperties(e,t),e}var Repository=function(){function e(r){var t,n,a,u,o=this,i=r.eventStore,s=r.aggregate;_classCallCheck(this,e),Object.defineProperty(this,"get",{configurable:!0,enumerable:!0,writable:!0,value:(t=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,o.eventStore.getEvents(r);case 2:if((t=e.sent).length){e.next=5;break}return e.abrupt("return",null);case 5:return e.abrupt("return",new o.Ctor(r,t));case 6:case"end":return e.stop()}},e,this)})),function(e){return t.apply(this,arguments)})}),Object.defineProperty(this,"create",{configurable:!0,enumerable:!0,writable:!0,value:function(e){return new o.Ctor(e||(0,_v.default)())}}),Object.defineProperty(this,"record",{configurable:!0,enumerable:!0,writable:!0,value:(n=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,o.eventStore.record(r);case 2:return t=e.sent,e.next=5,o.notifySubscribers(t);case 5:case"end":return e.stop()}},e,this)})),function(e){return n.apply(this,arguments)})}),Object.defineProperty(this,"notifySubscribers",{configurable:!0,enumerable:!0,writable:!0,value:(a=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,Promise.all(o.subscriptions.map(function(){var e=_asyncToGenerator(regeneratorRuntime.mark(function e(t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,t(r);case 3:return e.abrupt("return",e.sent);case 6:e.prev=6,e.t0=e.catch(0),console.log(e.t0);case 9:case"end":return e.stop()}},e,this,[[0,6]])}));return function(r){return e.apply(this,arguments)}}()));case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)})),function(e){return a.apply(this,arguments)})}),Object.defineProperty(this,"getEvents",{configurable:!0,enumerable:!0,writable:!0,value:(u=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,o.eventStore.getEvents(r);case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)})),function(e){return u.apply(this,arguments)})}),this.eventStore=i,this.Ctor=s,this.subscriptions=[]}return _createClass(e,[{key:"subscribe",value:function(e){this.subscriptions.push(e)}}]),e}();exports.default=Repository;
 //# sourceMappingURL=maps/repository.js.map
