@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function factory(e,r){return function(){return new BasicCommandHandler(e,r)}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=exports.EVENT=exports.COMMAND=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_extends2=require("babel-runtime/helpers/extends"),_extends3=_interopRequireDefault(_extends2),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_getPrototypeOf=require("babel-runtime/core-js/object/get-prototype-of"),_getPrototypeOf2=_interopRequireDefault(_getPrototypeOf),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_possibleConstructorReturn2=require("babel-runtime/helpers/possibleConstructorReturn"),_possibleConstructorReturn3=_interopRequireDefault(_possibleConstructorReturn2),_inherits2=require("babel-runtime/helpers/inherits"),_inherits3=_interopRequireDefault(_inherits2),_symbol=require("babel-runtime/core-js/symbol"),_symbol2=_interopRequireDefault(_symbol);exports.factory=factory;var _aggregateCommandHandler=require("./aggregateCommandHandler"),_aggregateCommandHandler2=_interopRequireDefault(_aggregateCommandHandler),COMMAND=exports.COMMAND=(0,_symbol2.default)("reference to command class"),EVENT=exports.EVENT=(0,_symbol2.default)("reference to event class"),BasicCommandHandler=function(e){function r(e,t){var n=this;(0,_classCallCheck3.default)(this,r);var a=(0,_possibleConstructorReturn3.default)(this,(r.__proto__||(0,_getPrototypeOf2.default)(r)).call(this,e));return a.execute=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",new a[EVENT]((0,_extends3.default)({},r)));case 1:case"end":return e.stop()}},e,n)}));return function(r){return e.apply(this,arguments)}}(),a[EVENT]=t,a}return(0,_inherits3.default)(r,e),r}(_aggregateCommandHandler2.default);exports.default=BasicCommandHandler;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.factory=factory,exports.default=exports.EVENT=exports.COMMAND=void 0;var _aggregateCommandHandler=_interopRequireDefault(require("./aggregateCommandHandler"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _typeof(e){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function _extends(){return(_extends=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e}).apply(this,arguments)}function _asyncToGenerator(e){return function(){var t=this,r=arguments;return new Promise(function(n,o){var a=e.apply(t,r);function i(e,t){try{var r=a[e](t),i=r.value}catch(e){return void o(e)}r.done?n(i):Promise.resolve(i).then(u,s)}function u(e){i("next",e)}function s(e){i("throw",e)}u()})}}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(e,t){if(t&&("object"===_typeof(t)||"function"==typeof t))return t;if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function _inherits(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}function _assertThisInitialized(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}var COMMAND=Symbol("reference to command class");exports.COMMAND=COMMAND;var EVENT=Symbol("reference to event class");exports.EVENT=EVENT;var BasicCommandHandler=function(e){function t(e,r){var n,o;return _classCallCheck(this,t),n=_possibleConstructorReturn(this,(t.__proto__||Object.getPrototypeOf(t)).call(this,e)),Object.defineProperty(_assertThisInitialized(n),"execute",{configurable:!0,enumerable:!0,writable:!0,value:(o=_asyncToGenerator(regeneratorRuntime.mark(function e(t){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",new n[EVENT](_extends({},t)));case 1:case"end":return e.stop()}},e,this)})),function(e){return o.apply(this,arguments)})}),n[EVENT]=r,n}return _inherits(t,_aggregateCommandHandler.default),t}();function factory(e,t){return function(){return new BasicCommandHandler(e,t)}}exports.default=BasicCommandHandler;
 //# sourceMappingURL=maps/basicCommandHandler.js.map
