@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function asArray(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return Array.isArray(e)?e:[e]}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),CommandHandler=function e(r){var t=this;(0,_classCallCheck3.default)(this,e),this.handle=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){var n,a;return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return n=new t.Command(r),e.next=3,n.validate();case 3:return e.next=5,t.authorize(n);case 5:return e.t0=asArray,e.next=8,t.execute(n);case 8:return e.t1=e.sent,a=(0,e.t0)(e.t1),a.forEach(function(e){e.actor=n.$identity&&n.$identity.userId,e.position=n.$position}),e.abrupt("return",a);case 12:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.execute=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.authorize=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.handleDeliveryError=function(){var e=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(r){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,t)}));return function(r){return e.apply(this,arguments)}}(),this.Command=r};exports.default=CommandHandler,module.exports=exports.default;
+"use strict";function _asyncToGenerator(e){return function(){var r=this,t=arguments;return new Promise(function(n,a){var o=e.apply(r,t);function i(e,r){try{var t=o[e](r),i=t.value}catch(e){return void a(e)}t.done?n(i):Promise.resolve(i).then(u,c)}function u(e){i("next",e)}function c(e){i("throw",e)}u()})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var CommandHandler=function e(r){var t,n,a,o,i=this;_classCallCheck(this,e),Object.defineProperty(this,"handle",{configurable:!0,enumerable:!0,writable:!0,value:(t=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=new i.Command(r),e.next=3,t.validate();case 3:return e.next=5,i.authorize(t);case 5:return e.t0=asArray,e.next=8,i.execute(t);case 8:return e.t1=e.sent,(n=(0,e.t0)(e.t1)).forEach(function(e){e.actor=t.$identity&&t.$identity.userId,e.position=t.$position}),e.abrupt("return",n);case 12:case"end":return e.stop()}},e,this)})),function(e){return t.apply(this,arguments)})}),Object.defineProperty(this,"execute",{configurable:!0,enumerable:!0,writable:!0,value:(n=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,this)})),function(e){return n.apply(this,arguments)})}),Object.defineProperty(this,"authorize",{configurable:!0,enumerable:!0,writable:!0,value:(a=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",!0);case 1:case"end":return e.stop()}},e,this)})),function(e){return a.apply(this,arguments)})}),Object.defineProperty(this,"handleDeliveryError",{configurable:!0,enumerable:!0,writable:!0,value:(o=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:case"end":return e.stop()}},e,this)})),function(e){return o.apply(this,arguments)})}),this.Command=r};function asArray(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return Array.isArray(e)?e:[e]}exports.default=CommandHandler;
 //# sourceMappingURL=maps/commandHandler.js.map
