@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),CommandHandlerError=function e(r){var l=r.error,a=r.handler,t=r.aggregate;(0,_classCallCheck3.default)(this,e),this.error=l,this.handler=a,this.aggregate=t};exports.default=CommandHandlerError,module.exports=exports.default;
+"use strict";function _classCallCheck(r,e){if(!(r instanceof e))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var CommandHandlerError=function r(e){var a=e.error,t=e.handler,o=e.aggregate;_classCallCheck(this,r),this.error=a,this.handler=t,this.aggregate=o};exports.default=CommandHandlerError;
 //# sourceMappingURL=../maps/commandScheduling/commandHandlerError.js.map
