@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),CommandScheduleTrigger=function e(r,t,a){var n=this;(0,_classCallCheck3.default)(this,e),this.start=function(){return setTimeout(n.onTick,n.interval)},this.onTick=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function e(){return _regenerator2.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,n.commandScheduler.deliverDueCommands(n.clock.now());case 2:setTimeout(n.onTick,n.interval);case 3:case"end":return e.stop()}},e,n)})),this.commandScheduler=r,this.clock=t,this.interval=a};exports.default=CommandScheduleTrigger,module.exports=exports.default;
+"use strict";function _asyncToGenerator(e){return function(){var n=this,t=arguments;return new Promise(function(r,o){var a=e.apply(n,t);function i(e,n){try{var t=a[e](n),i=t.value}catch(e){return void o(e)}t.done?r(i):Promise.resolve(i).then(c,u)}function c(e){i("next",e)}function u(e){i("throw",e)}c()})}}function _classCallCheck(e,n){if(!(e instanceof n))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var CommandScheduleTrigger=function e(n,t,r){var o,a=this;_classCallCheck(this,e),Object.defineProperty(this,"start",{configurable:!0,enumerable:!0,writable:!0,value:function(){return setTimeout(a.onTick,a.interval)}}),Object.defineProperty(this,"onTick",{configurable:!0,enumerable:!0,writable:!0,value:(o=_asyncToGenerator(regeneratorRuntime.mark(function e(){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,a.commandScheduler.deliverDueCommands(a.clock.now());case 2:setTimeout(a.onTick,a.interval);case 3:case"end":return e.stop()}},e,this)})),function(){return o.apply(this,arguments)})}),this.commandScheduler=n,this.clock=t,this.interval=r};exports.default=CommandScheduleTrigger;
 //# sourceMappingURL=../maps/commandScheduling/commandScheduleTrigger.js.map
