@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _promise=require("babel-runtime/core-js/promise"),_promise2=_interopRequireDefault(_promise),_regenerator=require("babel-runtime/regenerator"),_regenerator2=_interopRequireDefault(_regenerator),_asyncToGenerator2=require("babel-runtime/helpers/asyncToGenerator"),_asyncToGenerator3=_interopRequireDefault(_asyncToGenerator2),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_identity=require("../identity"),_identity2=_interopRequireDefault(_identity),_schedule=require("./schedule"),_schedule2=_interopRequireDefault(_schedule),_commandFailure=require("./commandFailure"),_commandFailure2=_interopRequireDefault(_commandFailure),CommandScheduler=function e(r){var t=r.store,n=r.clock,a=r.deliverer;(0,_classCallCheck3.default)(this,e),_initialiseProps.call(this),this.store=t,this.clock=n,this.deliverer=a},_initialiseProps=function(){var e=this;this.schedule=function(){var r=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function r(t){var n=t.service,a=t.target,u=t.clock,s=t.due,c=t.command;return _regenerator2.default.wrap(function(r){for(;;)switch(r.prev=r.next){case 0:return c.$identity=_identity2.default.system,c.$scheduler=new _schedule2.default({service:n,target:a,due:s,clock:u||e.clock,attempts:0}),r.next=4,e.store.push(c);case 4:case"end":return r.stop()}},r,e)}));return function(e){return r.apply(this,arguments)}}(),this.commandsDue=function(){var r=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function r(t){var n;return _regenerator2.default.wrap(function(r){for(;;)switch(r.prev=r.next){case 0:return r.next=2,e.store.commands();case 2:return n=r.sent,r.abrupt("return",n.filter(function(e){return e.$scheduler.isDue(t)}));case 4:case"end":return r.stop()}},r,e)}));return function(e){return r.apply(this,arguments)}}(),this.deliverDueCommands=function(){var r=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function r(t){var n;return _regenerator2.default.wrap(function(r){for(;;)switch(r.prev=r.next){case 0:return r.next=2,e.commandsDue(t);case 2:return n=r.sent,r.next=5,_promise2.default.all(n.map(function(){var r=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function r(t){return _regenerator2.default.wrap(function(r){for(;;)switch(r.prev=r.next){case 0:return r.next=2,e.deliver(t);case 2:return r.abrupt("return",r.sent);case 3:case"end":return r.stop()}},r,e)}));return function(e){return r.apply(this,arguments)}}()));case 5:case"end":return r.stop()}},r,e)}));return function(e){return r.apply(this,arguments)}}(),this.deliver=function(){var r=(0,_asyncToGenerator3.default)(_regenerator2.default.mark(function r(t){var n;return _regenerator2.default.wrap(function(r){for(;;)switch(r.prev=r.next){case 0:return r.prev=0,t.$scheduler.due=null,r.next=4,e.deliverer.deliver({service:t.$scheduler.service,target:t.$scheduler.target,command:t});case 4:if(t.$scheduler.attempts=t.$scheduler.attempts+1,!t.$scheduler.due){r.next=10;break}return r.next=8,e.store.retry(t);case 8:r.next=12;break;case 10:return r.next=12,e.store.complete(t);case 12:r.next=31;break;case 14:if(r.prev=14,r.t0=r.catch(0),n=new _commandFailure2.default({error:r.t0,command:t}),r.t0.handler){r.next=20;break}throw console.log("Unhandled error",r.t0),r.t0;case 20:return r.next=22,r.t0.handler.handleDeliveryError(n,r.t0.aggregate);case 22:if(!t.$scheduler.due){r.next=27;break}return r.next=25,e.store.retry(t);case 25:r.next=31;break;case 27:if(!n.cancelled){r.next=31;break}return console.log("retry cancelled"),r.next=31,e.store.complete(t);case 31:case"end":return r.stop()}},r,e,[[0,14]])}));return function(e){return r.apply(this,arguments)}}()};exports.default=CommandScheduler,module.exports=exports.default;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _identity=_interopRequireDefault(require("../identity")),_schedule=_interopRequireDefault(require("./schedule")),_commandFailure=_interopRequireDefault(require("./commandFailure"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _asyncToGenerator(e){return function(){var r=this,t=arguments;return new Promise(function(n,a){var u=e.apply(r,t);function c(e,r){try{var t=u[e](r),c=t.value}catch(e){return void a(e)}t.done?n(c):Promise.resolve(c).then(i,o)}function i(e){c("next",e)}function o(e){c("throw",e)}i()})}}function _classCallCheck(e,r){if(!(e instanceof r))throw new TypeError("Cannot call a class as a function")}var CommandScheduler=function e(r){var t=r.store,n=r.clock,a=r.deliverer;_classCallCheck(this,e),_initialiseProps.call(this),this.store=t,this.clock=n,this.deliverer=a};exports.default=CommandScheduler;var _initialiseProps=function(){var e,r,t,n,a=this;Object.defineProperty(this,"schedule",{configurable:!0,enumerable:!0,writable:!0,value:(e=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t,n,u,c,i;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t=r.service,n=r.target,u=r.clock,c=r.due,(i=r.command).$identity=_identity.default.system,i.$scheduler=new _schedule.default({service:t,target:n,due:c,clock:u||a.clock,attempts:0}),e.next=5,a.store.push(i);case 5:case"end":return e.stop()}},e,this)})),function(r){return e.apply(this,arguments)})}),Object.defineProperty(this,"commandsDue",{configurable:!0,enumerable:!0,writable:!0,value:(r=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,a.store.commands();case 2:return t=e.sent,e.abrupt("return",t.filter(function(e){return e.$scheduler.isDue(r)}));case 4:case"end":return e.stop()}},e,this)})),function(e){return r.apply(this,arguments)})}),Object.defineProperty(this,"deliverDueCommands",{configurable:!0,enumerable:!0,writable:!0,value:(t=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,a.commandsDue(r);case 2:return t=e.sent,e.next=5,Promise.all(t.map(function(){var e=_asyncToGenerator(regeneratorRuntime.mark(function e(r){return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,a.deliver(r);case 2:return e.abrupt("return",e.sent);case 3:case"end":return e.stop()}},e,this)}));return function(r){return e.apply(this,arguments)}}()));case 5:case"end":return e.stop()}},e,this)})),function(e){return t.apply(this,arguments)})}),Object.defineProperty(this,"deliver",{configurable:!0,enumerable:!0,writable:!0,value:(n=_asyncToGenerator(regeneratorRuntime.mark(function e(r){var t;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,r.$scheduler.due=null,e.next=4,a.deliverer.deliver({service:r.$scheduler.service,target:r.$scheduler.target,command:r});case 4:if(r.$scheduler.attempts=r.$scheduler.attempts+1,!r.$scheduler.due){e.next=10;break}return e.next=8,a.store.retry(r);case 8:e.next=12;break;case 10:return e.next=12,a.store.complete(r);case 12:e.next=31;break;case 14:if(e.prev=14,e.t0=e.catch(0),t=new _commandFailure.default({error:e.t0,command:r}),e.t0.handler){e.next=20;break}throw console.log("Unhandled error",e.t0),e.t0;case 20:return e.next=22,e.t0.handler.handleDeliveryError(t,e.t0.aggregate);case 22:if(!r.$scheduler.due){e.next=27;break}return e.next=25,a.store.retry(r);case 25:e.next=31;break;case 27:if(!t.cancelled){e.next=31;break}return console.log("retry cancelled"),e.next=31,a.store.complete(r);case 31:case"end":return e.stop()}},e,this,[[0,14]])})),function(e){return n.apply(this,arguments)})})};
 //# sourceMappingURL=../maps/commandScheduling/commandScheduler.js.map
