@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),RealWorldClock=function e(){(0,_classCallCheck3.default)(this,e),this.now=function(){return new Date}};exports.default=RealWorldClock,module.exports=exports.default;
+"use strict";function _classCallCheck(e,l){if(!(e instanceof l))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var RealWorldClock=function e(){_classCallCheck(this,e),Object.defineProperty(this,"now",{configurable:!0,enumerable:!0,writable:!0,value:function(){return new Date}})};exports.default=RealWorldClock;
 //# sourceMappingURL=../maps/commandScheduling/realWorldClock.js.map
