@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),Schedule=function e(t){var s=this,l=t.id,r=t.service,a=t.target,c=t.created,i=t.due,u=t.attempts,d=t.clock;(0,_classCallCheck3.default)(this,e),this.isDue=function(){return s.due<=s.clock.now()},this.id=l,this.service=r,this.target=a,this.created=c,this.due=i,this.attempts=u,this.clock=d};exports.default=Schedule,module.exports=exports.default;
+"use strict";function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var Schedule=function e(t){var s=this,c=t.id,i=t.service,a=t.target,r=t.created,l=t.due,o=t.attempts,n=t.clock;_classCallCheck(this,e),Object.defineProperty(this,"isDue",{configurable:!0,enumerable:!0,writable:!0,value:function(){return s.due<=s.clock.now()}}),this.id=c,this.service=i,this.target=a,this.created=r,this.due=l,this.attempts=o,this.clock=n};exports.default=Schedule;
 //# sourceMappingURL=../maps/commandScheduling/schedule.js.map
