@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _getPrototypeOf=require("babel-runtime/core-js/object/get-prototype-of"),_getPrototypeOf2=_interopRequireDefault(_getPrototypeOf),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_possibleConstructorReturn2=require("babel-runtime/helpers/possibleConstructorReturn"),_possibleConstructorReturn3=_interopRequireDefault(_possibleConstructorReturn2),_inherits2=require("babel-runtime/helpers/inherits"),_inherits3=_interopRequireDefault(_inherits2),_aggregateCommandHandler=require("./aggregateCommandHandler"),_aggregateCommandHandler2=_interopRequireDefault(_aggregateCommandHandler),CreateAggregateCommandHandler=function(e){function r(){var e,t,a,o;(0,_classCallCheck3.default)(this,r);for(var l=arguments.length,n=Array(l),u=0;u<l;u++)n[u]=arguments[u];return t=a=(0,_possibleConstructorReturn3.default)(this,(e=r.__proto__||(0,_getPrototypeOf2.default)(r)).call.apply(e,[this].concat(n))),a.isCreateHandler=!0,o=t,(0,_possibleConstructorReturn3.default)(a,o)}return(0,_inherits3.default)(r,e),r}(_aggregateCommandHandler2.default);exports.default=CreateAggregateCommandHandler,module.exports=exports.default;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _aggregateCommandHandler=_interopRequireDefault(require("./aggregateCommandHandler"));function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}function _typeof(e){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(e,t){if(t&&("object"===_typeof(t)||"function"==typeof t))return t;if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}function _inherits(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}function _assertThisInitialized(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}var CreateAggregateCommandHandler=function(e){function t(){var e,r,n;_classCallCheck(this,t);for(var o=arguments.length,a=new Array(o),i=0;i<o;i++)a[i]=arguments[i];return _possibleConstructorReturn(n,(r=n=_possibleConstructorReturn(this,(e=t.__proto__||Object.getPrototypeOf(t)).call.apply(e,[this].concat(a))),Object.defineProperty(_assertThisInitialized(n),"isCreateHandler",{configurable:!0,enumerable:!0,writable:!0,value:!0}),r))}return _inherits(t,_aggregateCommandHandler.default),t}();exports.default=CreateAggregateCommandHandler;
 //# sourceMappingURL=maps/createAggregateCommandHandler.js.map
