@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),Identity=function e(){var s=this,t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};(0,_classCallCheck3.default)(this,e),this.isAnonymous=function(){return s.userId===e.anonymous.userId},this.isSystem=function(){return s.userId===e.system.userId},this.username=t.username,this.userId=t.userId,this.claims=t.claims||{},this.roles=t.claims&&t.claims.roles||[],this.version=t.version};Identity.anonymous=new Identity({userId:"anonymous"}),Identity.system=new Identity({userId:"system"}),exports.default=Identity,module.exports=exports.default;
+"use strict";function _classCallCheck(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var Identity=function e(){var t=this,s=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};_classCallCheck(this,e),Object.defineProperty(this,"isAnonymous",{configurable:!0,enumerable:!0,writable:!0,value:function(){return t.userId===e.anonymous.userId}}),Object.defineProperty(this,"isSystem",{configurable:!0,enumerable:!0,writable:!0,value:function(){return t.userId===e.system.userId}}),this.username=s.username,this.userId=s.userId,this.claims=s.claims||{},this.roles=s.claims&&s.claims.roles||[],this.version=s.version};exports.default=Identity,Object.defineProperty(Identity,"anonymous",{configurable:!0,enumerable:!0,writable:!0,value:new Identity({userId:"anonymous"})}),Object.defineProperty(Identity,"system",{configurable:!0,enumerable:!0,writable:!0,value:new Identity({userId:"system"})});
 //# sourceMappingURL=maps/identity.js.map
