@@ -1,2 +1,2 @@
-"use strict";function _interopRequireDefault(e){return e&&e.__esModule?e:{default:e}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _getPrototypeOf=require("babel-runtime/core-js/object/get-prototype-of"),_getPrototypeOf2=_interopRequireDefault(_getPrototypeOf),_classCallCheck2=require("babel-runtime/helpers/classCallCheck"),_classCallCheck3=_interopRequireDefault(_classCallCheck2),_possibleConstructorReturn2=require("babel-runtime/helpers/possibleConstructorReturn"),_possibleConstructorReturn3=_interopRequireDefault(_possibleConstructorReturn2),_inherits2=require("babel-runtime/helpers/inherits"),_inherits3=_interopRequireDefault(_inherits2),ValidationError=function(e){function r(e){var t=e.aggregate,o=e.command,s=e.message;(0,_classCallCheck3.default)(this,r);var l=(0,_possibleConstructorReturn3.default)(this,(r.__proto__||(0,_getPrototypeOf2.default)(r)).call(this,s));return l.name="ValidationError",l.aggregate=t,l.command=o,l.message=s,l}return(0,_inherits3.default)(r,e),r}(Error);exports.default=ValidationError,module.exports=exports.default;
+"use strict";function _typeof(t){return(_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function _classCallCheck(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(t,e){if(e&&("object"===_typeof(e)||"function"==typeof e))return e;if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function _inherits(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,enumerable:!1,writable:!0,configurable:!0}}),e&&(Object.setPrototypeOf?Object.setPrototypeOf(t,e):t.__proto__=e)}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _gPO=Object.getPrototypeOf||function(t){return t.__proto__},_sPO=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t},_construct="object"===_typeof(Reflect)&&Reflect.construct||function(t,e,o){var r,n=[null];return n.push.apply(n,e),r=t.bind.apply(t,n),_sPO(new r,o.prototype)},_cache="function"==typeof Map&&new Map;function _wrapNativeSuper(t){if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==_cache){if(_cache.has(t))return _cache.get(t);_cache.set(t,e)}function e(){}return e.prototype=Object.create(t.prototype,{constructor:{value:e,enumerable:!1,writeable:!0,configurable:!0}}),_sPO(e,_sPO(function(){return _construct(t,arguments,_gPO(this).constructor)},t))}var ValidationError=function(t){function e(t){var o,r=t.aggregate,n=t.command,c=t.message;return _classCallCheck(this,e),(o=_possibleConstructorReturn(this,(e.__proto__||Object.getPrototypeOf(e)).call(this,c))).name="ValidationError",o.aggregate=r,o.command=n,o.message=c,o}return _inherits(e,_wrapNativeSuper(Error)),e}();exports.default=ValidationError;
 //# sourceMappingURL=maps/validationError.js.map
